+++ utils/layer_from_netCDF.py
@@ -47,7 +47,7 @@ def _get_vec_lyr(sqlite_path, layer_name, geom_column='the_geom'):
 def get_or_create_flowline_layer(ds, output_path):
     if not os.path.exists(output_path) or not \
             contains_layer(output_path, FLOWLINES_LAYER_NAME):
-        ga = ds.gridadmin  # TODO: to implement
+        ga = ds.gridadmin
         from .gridadmin import QgisLinesOgrExporter
         exporter = QgisLinesOgrExporter('dont matter')
         exporter.driver = ogr.GetDriverByName('SQLite')
@@ -62,7 +62,7 @@ def get_or_create_flowline_layer(ds, output_path):
 def get_or_create_node_layer(ds, output_path):
     if not os.path.exists(output_path) or not \
             contains_layer(output_path, NODES_LAYER_NAME):
-        ga = ds.gridadmin  # TODO: to implement
+        ga = ds.gridadmin
         from .gridadmin import QgisNodesOgrExporter
         exporter = QgisNodesOgrExporter('dont matter')
         exporter.driver = ogr.GetDriverByName('SQLite')
@@ -75,16 +75,17 @@ def get_or_create_node_layer(ds, output_path):
 
 @disable_sqlite_synchronous
 def get_or_create_pumpline_layer(ds, output_path):
+    ga = ds.gridadmin
     if not os.path.exists(output_path) or not \
             contains_layer(output_path, PUMPLINES_LAYER_NAME):
-        ga = ds.gridadmin  # TODO: to implement
-        from .gridadmin import QgisPumpsOgrExporter
-        exporter = QgisPumpsOgrExporter(node_data=ga.nodes.data)
-        exporter.driver = ogr.GetDriverByName('SQLite')
-        exporter.save(
-            output_path, PUMPLINES_LAYER_NAME, ga.pumps.data,
-            ga.pumps.epsg_code, 4326)
-    return _get_vec_lyr(output_path, PUMPLINES_LAYER_NAME)
+        if ga.has_pumpstations:
+            from .gridadmin import QgisPumpsOgrExporter
+            exporter = QgisPumpsOgrExporter(node_data=ga.nodes.data)
+            exporter.driver = ogr.GetDriverByName('SQLite')
+            exporter.save(output_path, PUMPLINES_LAYER_NAME, ga.pumps.data,
+                          ga.pumps.epsg_code, 4326)
+    if ga.has_pumpstations:
+        return _get_vec_lyr(output_path, PUMPLINES_LAYER_NAME)
 
 
 def make_flowline_layer(ds, spatialite, progress_bar=None):