+++ CHANGES.rst
@@ -4,6 +4,28 @@ threedi-qgis-plugin changelog
 
 0.16 (unreleased)
 -----------------
+- try to show more often the object_name in graph widget (instead of 'N/A')
+
+- use gridadmin has_pumpstations in functions get_or_create_pumpline_layer and
+  available_subgrid_map_vars
+
+- Progressbar exporting to gridadmin.sqlite starts now at 0%
+- Add support for aggregate netcdf in NetcdfDataSourceGroundwater.
+
+- add leakage to subgrid_map variables
+
+- Add support for aggregate netcdf in NetcdfDataSourceGroundwater.
+
+- add leakage to subgrid_map variables
+
+- Update available vars methods using threedigrid for
+  NetcdfDataSourceGroundwater.
+
+- Implement ``available_aggregation_vars`` using threedigrid.
+
+- Bump threedigrid to 0.2.2.
+
+- Enable threedigrid get_timeseries for result and schematization layers
 
 - Disable the "Calculate statistics?" prompt if there are already csv files
   available.